ds2.09a15

+++ lib/secure/sefun/user_exists.c
@@ -3,4 +3,4 @@
 //      returns true if the player exists
 //      created by Descartes of Borg 27 july 1993
 
-int user_exists(string str) { return (int)master()->player_exists(str); }
+int user_exists(string str) { return master()->player_exists(str); }
